+++ qom/solvers/QCMSolver.py
@@ -6,7 +6,7 @@
 __name__    = 'qom.solvers.QCMSolver'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2021-01-04'
-__updated__ = '2021-01-04'
+__updated__ = '2021-01-05'
 
 # dependencies
 from typing import Union
@@ -59,14 +59,143 @@ class QCMSolver():
         self.modes = modes
         self.corrs = corrs
 
-    def get_sync_complete(self, i_mode_i, i_mode_j):
+    def get_discord(self, idx_mode_i, idx_mode_j):
+        """Function to calculate Gaussian quantum discord between two modes given the correlation matrix of their quadratures.
+
+        Parameters
+        ----------
+        idx_mode_i : int
+            Index of ith mode.
+        idx_mode_j : int
+            Index of jth mode.
+
+        Returns
+        -------
+        D_G : float
+            Gaussian quantum discord.
+        """
+
+        # symplectic invariants
+        I_1, I_2, I_3, I_4 = self.__get_invariants(idx_mode_i, idx_mode_j)
+
+        try:
+            # sum of symplectic invariants
+            sigma = I_1 + I_2 + 2 * I_3
+            # symplectic eigenvalues
+            mu_plus = 1 / np.sqrt(2) * np.sqrt(sigma + np.sqrt(sigma**2 - 4 * I_4))
+            mu_minus = 1 / np.sqrt(2) * np.sqrt(sigma - np.sqrt(sigma**2 - 4 * I_4))
+        except ValueError:
+            logger.warning('Argument of sqrt is negative.\n')
+            return 0
+            
+        # quantum discord value
+        return self.__get_f(np.sqrt(I_2)) - self.__get_f(mu_plus) - self.__get_f(mu_minus) + self.__get_f(np.sqrt(self.__get_W(I_1, I_2, I_3, I_4)))
+
+    def __get_invariants(self, idx_mode_i, idx_mode_j):
+        """Helper function to calculate symplectic invariants for two modes given the correlation matrix of their quadratures.
+
+        Parameters
+        ----------
+        idx_mode_i : int
+            Index of ith mode.
+        idx_mode_j : int
+            Index of jth mode.
+
+        Returns
+        -------
+        invariants : list
+            Symplectic invariants.
+        """
+
+        # frequently used variables
+        pos_i = 2 * idx_mode_i
+        pos_j = 2 * idx_mode_j
+
+        # correlation matrix of ith mode
+        A = np.matrix([ [   self.corrs[pos_i][pos_i],     self.corrs[pos_i][pos_i + 1]      ],
+                        [   self.corrs[pos_i + 1][pos_i], self.corrs[pos_i + 1][pos_i + 1]  ]   ])
+        # correlation matrix of jth mode
+        B = np.matrix([ [   self.corrs[pos_j][pos_j],     self.corrs[pos_j][pos_j + 1]      ],
+                        [   self.corrs[pos_j][pos_j + 1], self.corrs[pos_j + 1][pos_j + 1]  ]   ])
+        # correlation matrix of intermodes
+        C = np.matrix([ [   self.corrs[pos_i][pos_j],     self.corrs[pos_i][pos_j + 1]      ],
+                        [   self.corrs[pos_i + 1][pos_j], self.corrs[pos_i + 1][pos_j + 1]  ]   ])
+        # correlation matrix of the modes
+        self.corrs_modes = np.matrix([[   self.corrs[pos_i][pos_i],     self.corrs[pos_i][pos_i + 1],     self.corrs[pos_i][pos_j],     self.corrs[pos_i][pos_j + 1]      ],
+                                    [   self.corrs[pos_i + 1][pos_i], self.corrs[pos_i + 1][pos_i + 1], self.corrs[pos_i + 1][pos_j], self.corrs[pos_i + 1][pos_j + 1]  ],
+                                    [   self.corrs[pos_j][pos_i],     self.corrs[pos_j][pos_i + 1],     self.corrs[pos_j][pos_j],     self.corrs[pos_j][pos_j + 1]      ],
+                                    [   self.corrs[pos_j + 1][pos_i], self.corrs[pos_j + 1][pos_i + 1], self.corrs[pos_j + 1][pos_j], self.corrs[pos_j + 1][pos_j + 1]  ]   ])
+
+        # symplectic invariants
+        return [np.linalg.det(A), np.linalg.det(B), np.linalg.det(C), np.linalg.det(self.corrs_modes)]
+
+    def __get_f(self, x):
+        """Helper function for quantum discord calculation."""
+
+        try:
+            return (x + 1 / 2) * np.log10(x + 1 / 2) - (x - 1 / 2) * np.log10(x - 1 / 2)
+        except ValueError:
+            logger.warning('Argument of log in f function is non-positive\n')
+            return 0
+
+    def __get_W(self, I_1, I_2, I_3, I_4):
+        """Helper function for quantum discord calculation."""
+
+        try: 
+            if 4 * (I_1 * I_2 - I_4)**2 / (I_1 + 4 * I_4) / (1 + 4 * I_2) / I_3**2 <= 1.0:
+                return ((2 * abs(I_3) + np.sqrt(4 * I_3**2 + (4 * I_2 - 1) * (4 * I_4 - I_1))) / (4 * I_2 - 1))**2
+            return (I_1 * I_2 + I_4 - I_3**2 - np.sqrt((I_1 * I_2 + I_4 - I_3**2)**2 - 4 * I_1 * I_2 * I_4)) / 2 / I_2
+        except ZeroDivisionError:
+            logger.warning('Division by zero encountered in W function\n')
+            return 0
+
+    def get_entan(self, idx_mode_i, idx_mode_j):
+        """Function to calculate quantum entanglement via logarithmic negativity between two modes given the correlation matrix of their quadratures.
+
+        Parameters
+        ----------
+        idx_mode_i : int
+            Index of ith mode.
+        idx_mode_j : int
+            Index of jth mode.
+
+        Returns
+        -------
+        E_N : float
+            Quantum entanglement value using logarithmic negativity.
+        """
+
+        # symplectic invariants
+        I_1, I_2, I_3, I_4 = self.__get_invariants(idx_mode_i, idx_mode_j)
+        
+        try:
+            # sum of symplectic invariants after positive partial transpose
+            sigma = I_1 + I_2 - 2*I_3
+            # smallest symplectic eigenvalue
+            mu_minus = 1 / np.sqrt(2) * np.sqrt(sigma - np.sqrt(sigma**2 - 4 * I_4))
+        except ValueError:
+            logger.warning('Argument of sqrt is negative\n')
+            E_N = 0
+
+        try:
+            # quantum entanglement value using logarithmic negativity
+            E_N =  -1 * (np.log(2 * mu_minus))
+        except ValueError:
+            logger.warning('Argument of ln is non-positive\n')
+            E_N = 0
+        else:
+            E_N = max(0, E_N)
+        
+        return E_N
+
+    def get_sync_complete(self, idx_mode_i, idx_mode_j):
         """Method to obtain the quantum complete synchronization measure between two modes.
 
         Parameters
         ----------
-        i_mode_i : int
+        idx_mode_i : int
             Index of ith mode.
-        i_mode_j : int
+        idx_mode_j : int
             Index of jth mode.
 
         Returns
@@ -76,8 +205,8 @@ class QCMSolver():
         """
 
         # frequently used variables
-        pos_i = 2 * i_mode_i
-        pos_j = 2 * i_mode_j
+        pos_i = 2 * idx_mode_i
+        pos_j = 2 * idx_mode_j
 
         # square difference between position quadratures
         q_minus_2 = 0.5 * (self.corrs[pos_i][pos_i] + self.corrs[pos_j][pos_j] - 2 * self.corrs[pos_i][pos_j])
@@ -91,14 +220,14 @@ class QCMSolver():
             logger.warning('Division by zero encountered in W function\n')
             return 0
 
-    def get_sync_phase(self, i_mode_i, i_mode_j):
+    def get_sync_phase(self, idx_mode_i, idx_mode_j):
         """Method to obtain the quantum phase synchronization measure between two modes.
 
         Parameters
         ----------
-        i_mode_i : int
+        idx_mode_i : int
             Index of ith mode.
-        i_mode_j : int
+        idx_mode_j : int
             Index of jth mode.
 
         Returns
@@ -108,12 +237,12 @@ class QCMSolver():
         """
 
         # frequently used variables
-        pos_i = 2 * i_mode_i
-        pos_j = 2 * i_mode_j
+        pos_i = 2 * idx_mode_i
+        pos_j = 2 * idx_mode_j
 
         # arguments
-        arg_i = np.angle(self.modes[i_mode_i])
-        arg_j = np.angle(self.modes[i_mode_j])
+        arg_i = np.angle(self.modes[idx_mode_i])
+        arg_j = np.angle(self.modes[idx_mode_j])
         
         # transformation for ith mode momentum quadrature
         p_i_prime_2 = (np.sin(arg_i))**2 * self.corrs[pos_i][pos_i] - (np.sin(arg_i)) * (np.cos(arg_i)) * self.corrs[pos_i][pos_i + 1] - (np.cos(arg_i)) * (np.sin(arg_i)) * self.corrs[pos_i + 1][pos_i] + (np.cos(arg_i))**2 * self.corrs[pos_i + 1][pos_i + 1]